Re-enable CH patch for kvm async pf
The original issue is in
https://github.com/cloud-hypervisor/cloud-hypervisor/issues/2277. In our
workload, when start CH on 2 physical cores, there will be high overhead
of CH itself (although the load in guest is low), which has been
verified using perf kvm. Further profiling using perf kvm stat, there
are many many VMExit (most of them are HLT). The phenomenon is similar
in the above issue link. But CH has stated that the issue has been fixed
in Linux 5.13. But I have tested on Linux 6.12.17 and ubunut 24 (6.8.0),
still facing the same problems. Re-enable the patch of CH seems work.